Add a simple login end-point and call from the frontend
We use Django cookie/session based auth for the API and Django Ninja requires
to enable CSRF for this authentication method. To have hot reload for the
front-end code, we use the webpack dev server when developing locally, and
redirect all UI requests to this server. This redirect also sets the CSRF token
cookie, so that the API requests work correctly.

Use a simple view to render the bundled JS in production