Force user to login to view the Home page
This is a trivial change to the home page view to verify that the login
workflow is working correctly.

The logged in status and the data returned by the `/api/user` end-point are not
stored in a global store, and therefore the API calls are made multiple times
to the end-point, but this can be fixed in a follow-up commit.