api: Use constants from backend when creating Razorpay orders
Previously, we sent the amount and the membership duration from the frontend
when creating Razorpay orders. This would mean users could tweak the request
coming from the front-end and change the amount being charged and the duration
of the membership. This commit changes the API end-point code, to use the
constants from the backend for membership fees and start, end dates.

Closes #36