transaction: Add a management command to sync Razorpay transactions
We have a webhook to handle successfully completed transactions, but in case
the webhook is down, or some events are missed, we can rely on the manually
synced data. Also, the webhook currently only listens to payment.completed
events and handles them. We could potentially listen to other types of events
too. If the webhook resends events that were not handled by the webhook, may be
extending the list of events we listen to is a better idea.