waiver: Use a boolean for waiver validity in the DB
Using the signed_by or signed_at fields to keep track of whether a waiver has
been signed is overloading the use of those fields. When a membership is
invalidated, we don't want to lose that information about the waiver signing
user and time, until the next waiver has been signed. This commit adds a
boolean to keep track of whether a membership has an associated, valid waiver.