test_ui: Add an integration test for the login flow
We use Zulip as an email server for the test. Instead of sending emails to a
real mailserver, we send the emails to a Zulip stream and use the Python Zulip
client to get the login link from the email.  This feels hackish, but works
very well without us having to run our own mailserver or find other services.