payment: Prevent multiple group payments for same player in the UI
This commit changes the UI to explicitly make the user want to make another
payment, after one payment is successful. The player list is updated on a
successful payment already, and the player search dropdown doesn't allow
selection of players with a valid memberhsip. This commit prevents a second
payment with an already selected list of players in the UI.