registration: Allow registering players >= minAge in calendar year
This commit allows registering players who are at least `minAge` years old in
the calendar year, when the registration form is being filled. In a follow up
commit, they will be allowed to pay membership fee only if they are at least
minAge years old by the end of the membership/season year. If appearing as a
guest player for an event, they need to be old enough to participate.