membership: Add restrictions on memberships for players under 12
Players need to be at least 12 to be able to participate in events. Restrict
individual annual memberships and event memberships if a player is not old
enough to participate in an event or the season. For group memberships, we only
show a warning to the user to ensure that all the players are of valid age.