validate-roster: Change end-point to list only registrations
This is a preparatory commit to allow for players to view registrations from
their own team. This commit changes the API end-point to list just
registrations instead of showing the full event information along with the
registrations. This list can be filtered down based on the user hitting the
end-point to show only relevant registrations, in a follow-up commit.