players: Remove team_name and add teams from UC
Previously, we allowed players to enter team names as a text field, and
normalizing the team names was a pain. This commit changes teams to be a many
to many relation on players, and the data is populated using Ultimate Central
as the source of truth. Teams are populated during the import, when linking UC
profiles to Hub, and when new registrations are imported from UC.

Closes #70

Co-authored-by: Puneeth Chaganti <punchagan@muse-amuse.in>