group-membership: Implement debounced search for players
Previously, we searched for players as we typed in the search bar, but this was
disabled in the commit where we changed the group memberships page to be
searchable by team names. This commit adds back this search functionality
where the search happens as the user types, but is debounced to make the search
occur only if the user stops typing for a second or so.