import_uc_events: Use bulk create and update non unique fields
Previously, we used get_or_create and iterated over a loop of the tournaments
fetched from Ultimate Central. In this commit we switch over to using a bulk
create, which would make the save more efficient, and also make sure that any
updates from UC are updated in our database.