fetch: Refactor fetchUserData to use success and failure callbacks
Previously, we passed the setData and setLoggedIn functions and the function
called them with appropriate arguments. Now, we extract the logic into separate
success and failure callbacks and use them everywhere. The callbacks are now a
part of the store module, which allows us to add more store logic directly.