ui: Cleanup fetching of user data and initializing flowbite
Extracted duplicated code to fetch user data and initialize flowbite into a
reusable component + a Route wrapper that simplifies and makes it easier to
understand how and where user data is being fetched. This could be a
preparatory commit to switch to solid-query to fetch the user data, along with
the work in 994f1cd1a0c821e431f0073e6580b4db78524bd1.

See #84