markdown: Make styling rendered markdown similar to the site easier
The About page uses part of the Markdown from the README to render it's
content. But, the content is completely unstyled because Solid Markdown doesn't
allow adding classes to the rendered HTML. This commit adds a component that
allows for specifying class tags to use on each HTML tag to render the page
using tailwind's classes.