transaction: API end-point to check status of a PhonePe transaction
If the transaction is pending, the server automatically checks the status of
the transaction with PhonePe. The reconciliation is triggered from the client
side to keep things simple. In a follow-up commit, server-side reconciliation
can be setup as a cron job. We won't exactly be following the algorithm
recommended by PhonePe for reconciliation, but this should be good enough for
our use case, without needing a separate service for reconciliations, etc.