index: Include webpack generated index file with hashed filenames
Previously, we used a modified version of the frontend/index.html as the Django
template which is served in production. We used ManifestStaticFilesStorage to
make sure the bundle filename has a hash in it, but the value of that doesn't
seem to change with every change to the frontend -- probably because of the use
of lazy loading components.  And since the filenames aren't changed, it causes
issues with caching.

This commit instead configure webpack to generate filenames with contenthashes
in them, and instead of using a script to generate the Django template, we
directly include the html from the webpack generated HTML.

Closes #223