uc_registration: Use email_address instead of email_canonical
As far as I can tell, `email_canonical` removes the "dots" (.) in the email,
and lower cases it. But, when trying to match this canonical email address
against addresses in our DB, we don't find the emails since we don't
canonicalize in the same way in our DB. This change switches over to using a
lower cased `email_address` from the registration information, instead.